Split webcash protocol stuff to be separate from wallet.
