+++ wallet.h
@@ -5,66 +5,17 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #ifndef WALLET_H
+#define WALLET_H
 
-#include <stdint.h>
+#include "webcash.h"
 
 #include <mutex>
 #include <string>
 
-#include "crypto/sha256.h"
-#include "sqlite3.h"
-#include "support/allocators/secure.h"
-#include "uint256.h"
-
-#include "absl/strings/escaping.h"
-#include "absl/strings/string_view.h"
-
 #include "boost/filesystem.hpp"
 #include "boost/interprocess/sync/file_lock.hpp"
 
-struct Amount {
-    int64_t i64;
-
-    Amount() : i64(0) {}
-    Amount(int64_t _i64) : i64(_i64) {}
-
-    bool parse(const absl::string_view& str);
-};
-
-inline bool operator==(const Amount& lhs, const Amount& rhs) { return lhs.i64 == rhs.i64; }
-inline bool operator!=(const Amount& lhs, const Amount& rhs) { return lhs.i64 != rhs.i64; }
-
-inline bool operator<(const Amount& lhs, const Amount& rhs) { return lhs.i64 < rhs.i64; }
-inline bool operator<=(const Amount& lhs, const Amount& rhs) { return lhs.i64 <= rhs.i64; }
-inline bool operator>=(const Amount& lhs, const Amount& rhs) { return lhs.i64 >= rhs.i64; }
-inline bool operator>(const Amount& lhs, const Amount& rhs) { return lhs.i64 > rhs.i64; }
-
-inline Amount operator-(const Amount& lhs, const Amount& rhs) { return Amount(lhs.i64 - rhs.i64); }
-inline Amount operator+(const Amount& lhs, const Amount& rhs) { return Amount(lhs.i64 + rhs.i64); }
-
-std::string to_string(const Amount& amt);
-
-struct SecretWebcash {
-    SecureString sk;
-    Amount amount;
-};
-
-std::string to_string(const SecretWebcash& esk);
-
-struct PublicWebcash {
-    uint256 pk;
-    Amount amount;
-
-    PublicWebcash(const SecretWebcash& esk)
-        : amount(esk.amount)
-    {
-        CSHA256()
-            .Write((const unsigned char*)esk.sk.c_str(), esk.sk.size())
-            .Finalize(pk.data());
-    }
-};
-
-std::string to_string(const PublicWebcash& epk);
+#include "sqlite3.h"
 
 class Wallet {
 protected:
